+++ cmd/ebrelayer/contract/contract.go
@@ -12,6 +12,8 @@ import (
 	"strings"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
+
+	"github.com/cosmos/peggy/cmd/ebrelayer/txs"
 )
 
 // BridgeBankABI path to file containing BridgeBank smart contract ABI
@@ -20,9 +22,9 @@ const BridgeBankABI = "/abi/BridgeBank.abi"
 // CosmosBridgeABI path to file containing CosmosBridge smart contract ABI
 const CosmosBridgeABI = "/abi/CosmosBridge.abi"
 
-// LoadABI loads a smart contract as an abi.ABI
-func LoadABI(cosmosSupport bool) abi.ABI {
-
+// LoadABI : loads a smart contract as an abi.ABI
+func LoadABI(contractType txs.ContractRegistry) abi.ABI {
+	// Open the file containing BridgeBank contract's ABI
 	var (
 		_, b, _, _ = runtime.Caller(0)
 		dir        = filepath.Dir(b)
@@ -30,13 +32,14 @@ func LoadABI(cosmosSupport bool) abi.ABI {
 
 	var contractRaw []byte
 	var err error
-	switch cosmosSupport {
-	case true:
+
+	switch contractType {
+	case txs.CosmosBridge:
 		contractRaw, err = ioutil.ReadFile(dir + CosmosBridgeABI)
 		if err != nil {
 			panic(err)
 		}
-	case false:
+	case txs.BridgeBank:
 		contractRaw, err = ioutil.ReadFile(dir + BridgeBankABI)
 		if err != nil {
 			panic(err)