+++ cmd/ebrelayer/contract/contract_test.go
@@ -4,6 +4,8 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+
+	"github.com/cosmos/peggy/cmd/ebrelayer/txs"
 )
 
 // TestLoadABI test that contract containing named event is successfully loaded
@@ -12,7 +14,7 @@ func TestLoadABI(t *testing.T) {
 	// const AbiPath = "/cmd/ebrelayer/contract/abi/BridgeBank.abi"
 
 	//Get the ABI ready
-	abi := LoadABI(true)
+	abi := LoadABI(txs.BridgeBank)
 
 	require.NotNil(t, abi.Events["LogLock"])
 }