+++ cmd/ebrelayer/relayer/ethereum.go
@@ -19,28 +19,28 @@ import (
 	sdkContext "github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
 	ethereum "github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/ethclient"
 
 	"github.com/cosmos/peggy/cmd/ebrelayer/contract"
 	"github.com/cosmos/peggy/cmd/ebrelayer/events"
 	"github.com/cosmos/peggy/cmd/ebrelayer/txs"
 )
 
-// InitEthereumRelayer Starts an event listener on a specific Ethereum network, contract, and event
+// InitEthereumRelayer : Subscribes to events emitted by the deployed contracts
 func InitEthereumRelayer(
 	cdc *codec.Codec,
-	chainID string,
 	provider string,
-	contractAddress common.Address,
-	makeClaims bool,
+	registryContractAddress common.Address,
 	validatorName string,
 	validatorAddress sdk.ValAddress,
-	cliContext sdkContext.CLIContext,
-	rpcURL string,
+	cliCtx sdkContext.CLIContext,
+	txBldr authtypes.TxBuilder,
 	privateKey *ecdsa.PrivateKey,
 ) error {
 	// Start client with infura ropsten provider
@@ -49,99 +49,101 @@ func InitEthereumRelayer(
 		return err
 	}
 
-	fmt.Printf("\nStarted Ethereum websocket with provider: %s", provider)
-
-	var targetContract txs.ContractRegistry
-	var eventName string
-
-	// Load target contract and event name
-	switch makeClaims {
-	case true:
-		targetContract = txs.CosmosBridge
-		eventName = events.LogNewProphecyClaim.String()
-	case false:
-		targetContract = txs.BridgeBank
-		eventName = events.LogLock.String()
-	}
-	// Load our contract's ABI
-	contractABI := contract.LoadABI(makeClaims)
-
-	// Load unique event signature from the named event contained within the contract's ABI
-	eventSignature := contractABI.Events[eventName].Id().Hex()
-
 	clientChainID, err := client.NetworkID(context.Background())
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	// Get the specific contract's address (CosmosBridge or BridgeBank)
-	targetAddress, err := txs.GetAddressFromBridgeRegistry(client, contractAddress, targetContract)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// We need the target address in bytes[] for the query
-	query := ethereum.FilterQuery{
-		Addresses: []common.Address{targetAddress},
-	}
-
 	// We will check logs for new events
 	logs := make(chan types.Log)
 
-	// Filter by contract and event, write results to logs
-	sub, err := client.SubscribeFilterLogs(context.Background(), query, logs)
-	if err != nil {
-		return err
-	}
-	fmt.Printf("\nSubscribed to %v contract at address: %s\n", targetContract, targetAddress.Hex())
+	// Start BridgeBank contract subscription
+	bridgeBankAddress, subBridgeBank := startContractEventSub(logs, client, registryContractAddress, txs.BridgeBank)
+
+	// Start CosmosBridge contract subscription
+	cosmosBridgeAddress, subCosmosBridge := startContractEventSub(logs, client, registryContractAddress, txs.CosmosBridge)
+
+	// Load BridgeBank contract ABI and LogLock event signature
+	bridgeBankContractABI := contract.LoadABI(txs.BridgeBank)
+	eventLogLockSignature := bridgeBankContractABI.Events[events.LogLock.String()].Id().Hex()
+
+	// Load CosmosBridge contract ABI and LogNewProphecyClaim event signature
+	cosmosBridgeContractABI := contract.LoadABI(txs.CosmosBridge)
+	eventLogNewProphecyClaimSignature := cosmosBridgeContractABI.Events[events.LogNewProphecyClaim.String()].Id().Hex()
 
 	for {
 		select {
 		// Handle any errors
-		case err := <-sub.Err():
+		case err := <-subBridgeBank.Err():
+			log.Fatal(err)
+		case err := <-subCosmosBridge.Err():
 			log.Fatal(err)
 		// vLog is raw event data
 		case vLog := <-logs:
-			// Check if the event is a 'LogLock' event
-			if vLog.Topics[0].Hex() == eventSignature {
-				fmt.Println("\nWitnessed new event:", eventName)
-				fmt.Println("Block number:", vLog.BlockNumber)
-				fmt.Println("Tx hash:", vLog.TxHash.Hex())
+			fmt.Println("\nWitnessed new Tx...")
+			fmt.Println("Block number:", vLog.BlockNumber)
+			fmt.Println("Tx hash:", vLog.TxHash.Hex())
 
-				var err error
-				switch eventName {
-				case events.LogLock.String():
-					err = handleLogLockEvent(
-						clientChainID, contractAddress, contractABI, eventName, vLog, chainID,
-						cdc, validatorAddress, validatorName, cliContext, rpcURL,
-					)
-				case events.LogNewProphecyClaim.String():
-					err = handleLogNewProphecyClaimEvent(
-						contractABI, eventName, vLog, provider, contractAddress, privateKey,
-					)
-				}
+			var err error
+			switch vLog.Topics[0].Hex() {
+			case eventLogLockSignature:
+				err = handleLogLockEvent(
+					clientChainID, bridgeBankAddress, bridgeBankContractABI, events.LogLock.String(),
+					vLog, cdc, validatorAddress, validatorName, cliCtx, txBldr,
+				)
+			case eventLogNewProphecyClaimSignature:
+				err = handleLogNewProphecyClaimEvent(
+					cosmosBridgeContractABI, events.LogNewProphecyClaim.String(), vLog, provider, cosmosBridgeAddress, privateKey,
+				)
+			}
 
-				if err != nil {
-					log.Fatal(err)
-				}
+			if err != nil {
+				log.Fatal(err)
 			}
 		}
 	}
 }
 
-// handleLogLockEvent unpacks a LogLock event, converts it to a ProphecyClaim, and relays a tx to Cosmos
+// startContractEventSub : starts an event subscription on the specified Peggy contract
+func startContractEventSub(
+	logs chan types.Log,
+	client *ethclient.Client,
+	registryAddress common.Address,
+	contractName txs.ContractRegistry,
+) (common.Address, ethereum.Subscription) {
+	// Get the contract address for this subscription
+	subContractAddress, err := txs.GetAddressFromBridgeRegistry(client, registryAddress, contractName)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// We need the address in []bytes for the query
+	subQuery := ethereum.FilterQuery{
+		Addresses: []common.Address{subContractAddress},
+	}
+
+	// Start the contract subscription
+	sub, err := client.SubscribeFilterLogs(context.Background(), subQuery, logs)
+	if err != nil {
+		log.Fatal(err)
+	}
+	fmt.Printf("\nSubscribed to %v contract at address: %s\n", contractName, subContractAddress.Hex())
+
+	return subContractAddress, sub
+}
+
+// handleLogLockEvent : unpacks a LogLock event, converts it to a ProphecyClaim, and relays a tx to Cosmos
 func handleLogLockEvent(
 	clientChainID *big.Int,
 	contractAddress common.Address,
 	contractABI abi.ABI,
 	eventName string,
 	log types.Log,
-	chainID string,
 	cdc *codec.Codec,
 	validatorAddress sdk.ValAddress,
 	validatorName string,
 	cliContext sdkContext.CLIContext,
-	rpcURL string,
+	txBldr authtypes.TxBuilder,
 ) error {
 	// Unpack the LogLock event using its unique event signature from the contract's ABI
 	event := events.UnpackLogLock(clientChainID, contractAddress.Hex(), contractABI, eventName, log.Data)
@@ -157,7 +159,7 @@ func handleLogLockEvent(
 
 	// Initiate the relay
 	return txs.RelayLockToCosmos(
-		chainID, cdc, validatorAddress, validatorName, cliContext, &prophecyClaim, rpcURL,
+		cdc, validatorName, &prophecyClaim, cliContext, txBldr,
 	)
 }